+++ src/main/java/com/nhom10/broadstore/services/PasswordHash.java
@@ -1,0 +1,81 @@
+package com.nhom10.broadstore.services;
+
+
+//package com.nhom10.broadstore.service;
+
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.PBEKeySpec;
+import java.math.BigInteger;
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
+import java.security.spec.InvalidKeySpecException;
+
+public class PasswordHash {
+    public static final String PBKDF2_ALGORITHM = "PBKDF2WithHmacSHA1";
+
+    // The following constants may be changed without breaking existing hashes.
+    public static final int SALT_BYTES = 24;
+    public static final int HASH_BYTES = 24;
+    public static final int PBKDF2_ITERATIONS = 1000;
+
+    public static String createHash(String password)
+            throws NoSuchAlgorithmException, InvalidKeySpecException {
+        return createHash(password.toCharArray());
+    }
+
+    public static String createHash(char[] password)
+            throws NoSuchAlgorithmException, InvalidKeySpecException {
+        // Generate a random salt
+        SecureRandom random = new SecureRandom();
+        byte[] salt = new byte[SALT_BYTES];
+
+        // Hash the password
+        byte[] hash = pbkdf2(password, salt, PBKDF2_ITERATIONS, HASH_BYTES);
+        // format iterations:salt:hash
+        return toHex(hash);
+    }
+
+    public static boolean validatePassword(String password, String goodHash)
+            throws NoSuchAlgorithmException, InvalidKeySpecException {
+        return validatePassword(password.toCharArray(), goodHash);
+    }
+
+    public static boolean validatePassword(char[] password, String goodHash)
+            throws NoSuchAlgorithmException, InvalidKeySpecException {
+
+        String testHash = createHash(password);
+        return goodHash.equals(testHash);
+    }
+
+    private static boolean slowEquals(byte[] a, byte[] b) {
+        int diff = a.length ^ b.length;
+        for (int i = 0; i < a.length && i < b.length; i++)
+            diff |= a[i] ^ b[i];
+        return diff == 0;
+    }
+
+    private static byte[] pbkdf2(char[] password, byte[] salt, int iterations, int bytes)
+            throws NoSuchAlgorithmException, InvalidKeySpecException {
+        PBEKeySpec spec = new PBEKeySpec(password, salt, iterations, bytes * 8);
+        SecretKeyFactory skf = SecretKeyFactory.getInstance(PBKDF2_ALGORITHM);
+        return skf.generateSecret(spec).getEncoded();
+    }
+
+    private static byte[] fromHex(String hex) {
+        byte[] binary = new byte[hex.length() / 2];
+        for (int i = 0; i < binary.length; i++) {
+            binary[i] = (byte) Integer.parseInt(hex.substring(2 * i, 2 * i + 2), 16);
+        }
+        return binary;
+    }
+
+    private static String toHex(byte[] array) {
+        BigInteger bi = new BigInteger(1, array);
+        String hex = bi.toString(16);
+        int paddingLength = (array.length * 2) - hex.length();
+        if (paddingLength > 0)
+            return String.format("%0" + paddingLength + "d", 0) + hex;
+        else
+            return hex;
+    }
+}
